raspberry/shuttercontrol-service/requirements.txt
paho-mqtt
smbus